pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "cde"
version = "0.0.1"
description = "TPU/GPU iteration manager — build, submit, record, compare."
requires-python = ">=3.10"
dependencies = [
    "pyyaml >= 6.0",
    "jinja2 >= 3.0",
    "watchdog >= 3.0",
    "argcomplete >= 3.0",
]

[project.optional-dependencies]
dev = [
    "pytest >= 7.0",
]

[project.scripts]
cde = "cde.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cde = ["templates/*"]
